Move execution context into a seperate crate
