+++ executor/context/Cargo.toml
@@ -16,7 +16,7 @@
 # under the License.
 
 [package]
-name = "teaclave_executor"
+name = "teaclave_executor_context"
 version = "0.2.0"
 authors = ["Teaclave Contributors <dev@teaclave.apache.org>"]
 description = "Teaclave executor"
@@ -24,7 +24,7 @@ license = "Apache-2.0"
 edition = "2018"
 
 [lib]
-name = "teaclave_executor"
+name = "teaclave_executor_context"
 crate-type = ["staticlib", "rlib"]
 
 [features]
@@ -34,47 +34,16 @@ mesalock_sgx = [
   "teaclave_types/mesalock_sgx",
   "teaclave_crypto/mesalock_sgx",
   "teaclave_runtime/mesalock_sgx",
-  "teaclave_function/mesalock_sgx",
-  "teaclave_executor_context/mesalock_sgx",
 ]
 cov = ["sgx_cov"]
 enclave_unit_test = [
   "teaclave_test_utils/mesalock_sgx",
-  "teaclave_runtime/mesalock_sgx",
+  "teaclave_runtime/mesalock_sgx"
 ]
 
-# Enable builtin functions for the builtin executor
-
-full_builtin_function = [
-  "builtin_echo",
-  "builtin_face_detection",
-  "builtin_gbdt_predict",
-  "builtin_gbdt_train",
-  "builtin_logistic_regression_predict",
-  "builtin_logistic_regression_train",
-  "builtin_password_check",
-  "builtin_online_decrypt",
-  "builtin_ordered_set_intersect",
-  "builtin_principal_components_analysis",
-  "builtin_private_join_and_compute",
-  "builtin_rsa_sign",
-]
-
-builtin_echo = []
-builtin_face_detection = []
-builtin_gbdt_predict = []
-builtin_gbdt_train = []
-builtin_logistic_regression_predict = []
-builtin_logistic_regression_train = []
-builtin_password_check = []
-builtin_online_decrypt = []
-builtin_ordered_set_intersect = []
-builtin_principal_components_analysis = []
-builtin_private_join_and_compute = []
-builtin_rsa_sign = []
 
 [dependencies]
-log           = { version = "0.4.6", features = ["release_max_level_info"] }
+log           = { version = "0.4.6", features = ["release_max_level_debug"] }
 anyhow        = { version = "1.0.26" }
 serde_json    = { version = "1.0.39" }
 serde         = { version = "1.0.92", features = ["derive"] }
@@ -83,12 +52,10 @@ gbdt          = { version = "0.1.0", features = ["input", "enable_training"] }
 rusty-machine = { version = "0.5.4" }
 itertools     = { version = "0.8.0", default-features = false }
 
-teaclave_types      = { path = "../types" }
-teaclave_crypto     = { path = "../crypto" }
-teaclave_runtime    = { path = "../runtime", optional = true }
-teaclave_test_utils = { path = "../tests/utils", optional = true }
-teaclave_function   = { path = "../function" }
-teaclave_executor_context = { path = "./context" }
+teaclave_types      = { path = "../../types" }
+teaclave_crypto     = { path = "../../crypto" }
+teaclave_runtime    = { path = "../../runtime", optional = true }
+teaclave_test_utils = { path = "../../tests/utils", optional = true }
 
 sgx_cov       = { version = "1.1.3", optional = true }
 sgx_tstd      = { version = "1.1.3", features = ["net", "thread", "backtrace"], optional = true }
