+++ function/Cargo.toml
@@ -34,6 +34,7 @@ mesalock_sgx = [
   "teaclave_types/mesalock_sgx",
   "teaclave_crypto/mesalock_sgx",
   "teaclave_runtime/mesalock_sgx",
+  "teaclave_executor_context/mesalock_sgx",
 ]
 cov = ["sgx_cov"]
 enclave_unit_test = [
@@ -60,6 +61,7 @@ teaclave_types = { path = "../types" }
 teaclave_crypto = { path = "../crypto" }
 teaclave_runtime = { path = "../runtime", optional = true }
 teaclave_test_utils = { path = "../tests/utils", optional = true }
+teaclave_executor_context = { path = "../executor/context" }
 
 sgx_cov       = { version = "1.1.3", optional = true }
 sgx_tstd      = { version = "1.1.3", features = ["net", "thread", "backtrace"], optional = true }
